+++ src/xyz/oihana/schema/constants/traits/StatisticsTrait.php
@@ -2,7 +2,9 @@
 
 namespace xyz\oihana\schema\constants\traits;
 
+use xyz\oihana\schema\constants\traits\statistics\CustomerReceivablesTrait;
 use xyz\oihana\schema\constants\traits\statistics\CustomerStatisticsTrait;
+use xyz\oihana\schema\constants\traits\statistics\HasReceivablesTrait;
 use xyz\oihana\schema\constants\traits\statistics\HasTradingMeasuresTrait;
 use xyz\oihana\schema\constants\traits\statistics\HasUninvoicedTradeTrait;
 use xyz\oihana\schema\constants\traits\statistics\ObservationSeriesTrait;
@@ -20,12 +22,14 @@ use xyz\oihana\schema\constants\traits\statistics\StatisticsSummaryTrait;
  */
 trait StatisticsTrait
 {
-    use CustomerStatisticsTrait ,
-        HasTradingMeasuresTrait ,
-        HasUninvoicedTradeTrait ,
-        ObservationSeriesTrait  ,
-        SalesObjectivesTrait    ,
-        SellerStatisticsTrait   ,
-        StatisticsRecordTrait   ,
-        StatisticsSummaryTrait  ;
+    use CustomerReceivablesTrait ,
+        CustomerStatisticsTrait  ,
+        HasReceivablesTrait      ,
+        HasTradingMeasuresTrait  ,
+        HasUninvoicedTradeTrait  ,
+        ObservationSeriesTrait   ,
+        SalesObjectivesTrait     ,
+        SellerStatisticsTrait    ,
+        StatisticsRecordTrait    ,
+        StatisticsSummaryTrait   ;
 }